+++ georssy/__init__.py
@@ -1,3 +1,1 @@
-#from . import decoder
-#from .gml import gml_decoder
-#from .simple import simple_decoder
+from api import decode as GeoRssDecoder