+++ HISTORY.rst
@@ -4,6 +4,15 @@ Release History
 ---------------
 
 dev (2017-08-24)
-+++
+++++++++++++++++
+
+- Added GeoRSSEntry model class
+- Created "Tests" folder and moved Test.py into it
+- Added "decoder" functions into simple_decoder.py and gml_decoder.py. Changed decoder.py implementation
+- Added api.py module
+- Prepared for pip release
+
+dev (2017-08-24)
+++++++++++++++++
 
 - First Commit